Splitting SqlClientDiagnostic.netcore.cs into one file per class
